🍫 (#54) - move popup to popup component
