+++ nutella/src/components/MarkdownEditor/Row.tsx
@@ -3,14 +3,9 @@ import { MarkdownContext, Row as RowType } from "../../context/MarkdownCotext";
 import * as S from "./styles";
 import HandleSVG from "../../assets/icons/handle.svg";
 import Tag from "../../interface/Tag";
+import Popup from "./Popup";
 interface PropsType {
   data: RowType;
-}
-
-interface PopupItem {
-  title: string;
-  description: string[];
-  onClick: () => void;
 }
 
 const tagMap = new Map<string, Tag>()
@@ -65,64 +60,6 @@ const Row: FC<PropsType> = ({ data }) => {
     changeTab,
   } = useContext(MarkdownContext);
   const currentIndex = rows.findIndex((value) => value.id === id);
-
-  const popupItems = useMemo<PopupItem[]>(
-    () => [
-      {
-        title: "제목 1",
-        description: ["2.25rem | 굵음", "단축 : #"],
-        onClick: () => changeRowType(id, "h1"),
-      },
-      {
-        title: "제목 2",
-        description: ["1.75rem | 굵음", "단축 : ##"],
-        onClick: () => changeRowType(id, "h2"),
-      },
-      {
-        title: "제목 3",
-        description: ["1.5rem | 굵음", "단축 : ###"],
-        onClick: () => changeRowType(id, "h3"),
-      },
-      {
-        title: "제목 4",
-        description: ["1.25rem | 중간", "단축 : ####"],
-        onClick: () => changeRowType(id, "h4"),
-      },
-      {
-        title: "제목 5",
-        description: ["1.25rem | 중간", "단축 : #####"],
-        onClick: () => changeRowType(id, "h5"),
-      },
-      {
-        title: "제목 6",
-        description: ["1rem | 굵음", "단축 : ######"],
-        onClick: () => changeRowType(id, "h6"),
-      },
-      {
-        title: "인용",
-        description: ["1rem | 표준", "단축 : >"],
-        onClick: () => changeRowType(id, "blockquote"),
-      },
-      {
-        title: "순서 있는 목록",
-        description: ["1rem | 표준", "단축 : {아무 숫자}."],
-        onClick: () => changeRowType(id, "ol"),
-      },
-      {
-        title: "순서 없는 목록",
-        description: ["1rem | 표준", "단축 : -"],
-        onClick: () => changeRowType(id, "ul"),
-      },
-    ],
-    []
-  );
-
-  useEffect(() => {
-    if (refs.current) {
-      refs.current[currentIndex].focus();
-      refs.current[currentIndex].innerText = text;
-    }
-  }, [refs, type]);
 
   const onKeyDown = useCallback(
     (e: React.KeyboardEvent<HTMLDivElement>) => {
@@ -265,28 +202,18 @@ const Row: FC<PropsType> = ({ data }) => {
     [type, tab, renderRow, getStart]
   );
 
+  useEffect(() => {
+    if (refs.current) {
+      refs.current[currentIndex].focus();
+      refs.current[currentIndex].innerText = text;
+    }
+  }, [refs, type]);
+
   return (
     <S.RowContainer margin={`calc(${tab} * 1.2rem)`}>
       <S.Handle className="handle">
         <S.HandleIcon src={HandleSVG} />
-        <S.PopUp className="popup">
-          <S.PopupRowTitle>행 종류</S.PopupRowTitle>
-          <S.PopupRowOuter>
-            {popupItems.map((value, index) => {
-              const { title, description, onClick } = value;
-              return (
-                <S.PopupRowContainer onClick={onClick} key={index}>
-                  <S.PopupRowTitle>{title}</S.PopupRowTitle>
-                  <S.PopupRowDescription>
-                    {description.map((value) => (
-                      <div>{value}</div>
-                    ))}
-                  </S.PopupRowDescription>
-                </S.PopupRowContainer>
-              );
-            })}
-          </S.PopupRowOuter>
-        </S.PopUp>
+        <Popup id={id} />
       </S.Handle>
       {isList(type) ? renderListRow : renderRow}
     </S.RowContainer>
